src/jolle/requirements.txt
transit-python==0.8.302
requests==2.19.1
